added responsive login view
v0.1.0

This commit focused on responsive login view.

Compatibility:
Desktop
Tablet
Phone

+++ lib/features/layout/constant/constant_screen_sizes.dart
@@ -1,4 +1,4 @@
-const int phoneSize = 640;
-const int tabletSize = 641;
-const int tabletBreakPointSize = 1007;
-const int desktopSize = 1008;
+const double phoneSize = 640;
+const double tabletSize = 641;
+const double tabletBreakPointSize = 1007;
+const double desktopSize = 1008;
